% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqr_gweis.R
\name{bqr_gweis}
\alias{bqr_gweis}
\title{Perform GWEIS for binary outcome and quantitative exposure variables.}
\usage{
bqr_gweis(plink_path, dis_snp, qr_dis_cov, br_dis_phen, output_dir)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{qr_dis_cov}{File path for covariate data in the discovery dataset.}

\item{br_dis_phen}{File path for the phenotype data in the discovery dataset.}

\item{output_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.
}
\description{
This function performs genome-wide interaction studies (GWEIS) and processes the results
to generate files for downstream analysis.
}
