% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_binary1_merged.R
\name{gcim_b1_merged}
\alias{gcim_b1_merged}
\title{Perform regression analysis for genetic causality inference model(GCIM) with binary outcome}
\usage{
gcim_b1_merged(bp_tar_phen, verbose = TRUE)
}
\arguments{
\item{bp_tar_phen}{File path to the phenotype and PRS data (tab-delimited).
Must contain at least:
FID, IID, Outcome, Add_PRS, Int_PRS, Cov_PRS, Covariate_Pheno.
Additional columns will be treated as confounders.}

\item{verbose}{Logical; whether to print progress messages (default: TRUE).}
}
\value{
A list containing:
\item{model}{The fitted \code{glm} model object.}
\item{model_summary}{Summary of the fitted model.}
\item{merged_data}{Data frame used for model fitting.}
\item{formula}{The model formula used.}
\item{sample_size}{Number of observations used in the analysis.}
\item{variables}{Names of the variables in the model.}
\item{pseudo_r_squared}{Pseudo R-squared value of the model.}
\item{null_deviance}{Null deviance of the model.}
\item{residual_deviance}{Residual deviance of the model.}
\item{aic}{AIC of the model.}
}
\description{
This function performs logistic regression analysis for GCIM
with binary outcomes using the main effects of Polygenic Risk Scores (PRS) of the exposure.
It can read PRS files from temporary directories or use provided PRS data objects.
from own PRS data.
}
