% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbp_gweis.R
\name{gcim_bbp}
\alias{gcim_bbp}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_bbp(
  bp_tar_phen,
  bp_tar_cov,
  output_dir,
  Confounders,
  Additive_data,
  Interaction_data,
  Covariate_prs
)
}
\arguments{
\item{bp_tar_phen}{File path for the target phenotype data.}

\item{bp_tar_cov}{File path for the target covariate data.}

\item{Confounders}{Data frame of additional confounders.}

\item{Additive}{Scaled additive PRS values.}

\item{Interaction}{Scaled interaction PRS values.}

\item{Covariate}{Scaled covariate PRS values.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
