% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcim_q.R
\name{gcim_q}
\alias{gcim_q}
\title{Perform regression analysis for GCIM with quantitative outcome.}
\usage{
gcim_q(qp_tar_phen, qp_tar_cov, Add_PRS, Int_PRS, Cov_PRS, confounders = NULL)
}
\arguments{
\item{qp_tar_phen}{File path for the target phenotype data.}

\item{qp_tar_cov}{File path for the target covariate data.}

\item{Add_PRS}{File path for scaled additive PRS values.}

\item{Int_PRS}{File path for scaled interaction PRS values.}

\item{Cov_PRS}{File path for scaled covariate PRS values.}

\item{confounders}{Optional data frame of additional confounders (not used if covariates file contains confounders).}
}
\value{
Summary of the regression model.
}
\description{
Perform regression analysis for GCIM with quantitative outcome.
}
