% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbp_gweis.R
\name{gcim_qbp}
\alias{gcim_qbp}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_qbp(qp_tar_phen, bp_tar_cov, Add_PRS, Int_PRS, Cov_PRS, confounders)
}
\arguments{
\item{qp_tar_phen}{File path for the target phenotype data.}

\item{bp_tar_cov}{File path for the target covariate data.}

\item{confounders}{Data frame of additional confounders.}

\item{prs_add_scaled}{Scaled additive PRS values.}

\item{prs_int_scaled}{Scaled interaction PRS values.}

\item{prs_cov_scaled}{Scaled covariate PRS values.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
