% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbp_gweis.R
\name{qbp_gweis}
\alias{qbp_gweis}
\title{Perform GWEIS for quantitative outcome and binary exposure variables.}
\usage{
qbp_gweis(plink_path, dis_snp, qp_dis_phen, bp_dis_cov, output_dir)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{qp_dis_phen}{File path for the phenotype data in the discovery dataset.}

\item{bp_dis_cov}{File path for covariate data in the discovery dataset.}

\item{output_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.
}
\description{
This function performs genome-wide interaction studies (GWEIS) and processes the results
to generate files for downstream analysis.
}
