% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbr_gweis.R
\name{qbr_gweis}
\alias{qbr_gweis}
\title{Perform GWEIS for quantitative outcome and binary exposure variables for the reverse directions.}
\usage{
qbr_gweis(plink_path, dis_snp, br_dis_cov, qr_dis_phen, output_dir)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{br_dis_cov}{File path for covariate data in the discovery dataset.}

\item{qr_dis_phen}{File path for the phenotype data in the discovery dataset.}

\item{output_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.
}
\description{
This function performs genome-wide interaction studies (GWEIS) and processes the results
to generate files for downstream analysis.
}
