% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqp_gweis.R, R/qqp_gweis.R
\name{qp_gwas}
\alias{qp_gwas}
\title{Perform GWAS for covariates.}
\usage{
qp_gwas(plink_path, dis_snp, qp_dis_cov, output_dir)

qp_gwas(plink_path, dis_snp, qp_dis_cov, output_dir)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{qp_dis_cov}{File path for covariate data in the discovery dataset.}

\item{output_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.

None. Results are saved to files.
}
\description{
Perform GWAS for covariates.

Perform GWAS for covariates.
}
