% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbr_gweis.R, R/qqr_gweis.R
\name{qr_gwas}
\alias{qr_gwas}
\title{Perform GWAS for covariates.}
\usage{
qr_gwas(plink_path, dis_snp, qr_dis_phen, output_dir)

qr_gwas(plink_path, dis_snp, qr_dis_phen, output_dir)
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset.}

\item{qr_dis_phen}{File path for covariate data in the discovery dataset.}

\item{output_dir}{Directory to save output files.}
}
\value{
None. Results are saved to files.

None. Results are saved to files.
}
\description{
Perform GWAS for covariates.

Perform GWAS for covariates.
}
