% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbp_gweis.R
\name{gcim_bbp}
\alias{gcim_bbp}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_bbp(
  bp_tar_phen,
  bp_tar_cov,
  Additive,
  Interaction,
  Covariate,
  Confounders
)
}
\arguments{
\item{bp_tar_phen}{File path for the target phenotype data.}

\item{bp_tar_cov}{File path for the target covariate data.}

\item{prs_add}{Scaled additive PRS values.}

\item{prs_int}{Scaled interaction PRS values.}

\item{prs_cov}{Scaled covariate PRS values.}

\item{confounders}{Data frame of additional confounders.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
