% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqr_gweis.R
\name{gcim_bqr}
\alias{gcim_bqr}
\title{Perform Regression Analysis for GCIM.}
\usage{
gcim_bqr(qr_tar_cov, br_tar_phen, Add_PRS, Int_PRS, Cov_PRS, confounders)
}
\arguments{
\item{qr_tar_cov}{File path for the target covariate data.}

\item{br_tar_phen}{File path for the target phenotype data.}

\item{confounders}{Data frame of additional confounders.}

\item{prs_add_scaled}{Scaled additive PRS values.}

\item{prs_int_scaled}{Scaled interaction PRS values.}

\item{prs_cov_scaled}{Scaled covariate PRS values.}
}
\value{
Summary of the regression model.
}
\description{
Perform Regression Analysis for GCIM.
}
