% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbp_gweis.R
\name{qbp_gweis}
\alias{qbp_gweis}
\title{Perform GWEIS, GWAS, PRS computation, and regression analysis for the proposed direction of causations in quantitative outcome and binary exposure variable.}
\usage{
qbp_gweis()
}
\arguments{
\item{plink_path}{Path to the PLINK executable application.}

\item{dis_snp}{Prefix for binary files for the discovery dataset (common prefix for .fam, .bim, .bed files).}

\item{tar_snp}{Prefix for binary files for the target dataset (common prefix for .fam, .bim, .bed files).}

\item{qp_dis_phen}{File path for the phenotype data of the binary outcome in the discovery dataset.}

\item{qp_tar_phen}{File path for the phenotype data of the binary outcome in the target dataset.}

\item{bp_dis_cov}{File path for covariate data (exposure and confounders) in the discovery dataset for quatitative exposure.}

\item{bp_tar_cov}{File path for covariate data (exposure and confounders) in the target dataset for quatitative exposure.}

\item{output_dir}{Directory for saving output files.}

\item{confounders}{Covariates to include as confounders in the regression model.}
}
\value{
Summary of the regression model.
}
\description{
This function performs genome-wide interaction studies (GWEIS), genome-wide association studies (GWAS),
polygenic risk score (PRS) computation, and regression analysis to determine causal directions
for binary outcomes and binary exposure variables.
}
